feat: submenu Financeiro, config separada, boleto sem CPF manual
- MainListItems: Financeiro vira submenu (Faturas + Configurações)
- Nova página /financeiro-config com formulário de dados fiscais
- Faturas: remove formulário fiscal (agora em /financeiro-config)
- CheckoutPage: remove campo CPF/CNPJ (backend usa company.document)
- asaasCreateBoleto: remove validação cpfCnpj, usa dados da empresa;
  log do invoiceId para diagnóstico de "Invoice not found"
